Add some raw notes on the ARAM maps for some N-SPC/Kankichi-kun variants
These will be among my earlier test cases to try to adapt for displaying on a
page. The changelogs are provided so that the ARAM maps can be dynamically
generated from them: the intention is that this will simplify the process of
producing the documentation in the long run.

I have covered the following to begin with...
- N-SPC/Standard
-- Traced from Super Mario World and Pilotwings, which fall under
   N-SPC/Nintendo/Koji Kondo/Prototype (as version-wise they predate Standard).
- N-SPC/Nintendo/Koji Kondo
-- The notes are incomplete, particularly for the later games.
- N-SPC/Nintendo/Unique/F-Zero
-- Covers the beta and final versions.