Add David Whittaker's sound driver to the ARAM Map raw notes
This is the sound driver for which I have both the most complete set of notes on
hand and also the one which I released as a SNESLab article. Only the ARAM map
was not covered, and that was because I had not previously recorded it. I am
doing so now so that this sound driver can be the first that I can fully cover,
since its featuresets can vary wildly between builds, and it contains test cases
for both relocating memory locations and expansions of the number of striped
bytes allocated.

Currently, only the four earliest builds are done at this time. Batman: Revenge
of the Joker is next on my list, and it definitely contains some ARAM map
relocations. The version numbering is also a WIP, mostly carried over from the
SNESLab article. However, there are two differences: the first is that this
sound driver has an identical build that produces a different ARAM map for its
variable ARAM section that is not caused by code being put in between. Thus, it
must be counted. The only other case I can think of off of the top of my head is
in Hudson's sound driver. The second is that Kick Off is put earlier than
Krusty's Super Fun House: this only applies to the ARAM map. The reason for this
is the number of channels allocated to the music: Dream TV (Beta 2) allocates 5
channels for the music. Kick Off does the same thing, and the ARAM map between
the two corresponds with that. Krusty's Super Fun House, although it is an
earlier revision by code, allocates 6 channels for the music, thus having more
differences. Critically, though, it doesn't add any locations of its own through
its code differences: thus, most tracing except for the ARAM map would use
Krusty's Super Fun House first.